bling-dashboard/requirements.txt
streamlit>=1.32
pandas>=2.0
duckdb>=0.10
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
plotly>=5.18
